Refactored name to GetBotAccessTokenService
